Created getUserByID method, also changed ID back to an int

+++ src/main/java/com/jasirid/demo/service/UserService.java
@@ -5,5 +5,5 @@ import com.jasirid.demo.dto.UserDTO;
 public interface UserService {
     UserDTO createUser(UserDTO userDTO);
 
-    UserDTO getUserByID(Long userID);
+    UserDTO getUserByID(int userID);
 }
